chore: Fix merge script and add api-logs, api-events packages
