+++ sandbox-tools/merge-repos/src/config.ts
@@ -129,7 +129,9 @@ export const foldersToMerge: IMergePackageDetail[] = [
     { name: "@opentelemetry/instrumentation-user-interaction", destPath: "pkgs/instrumentations/web/user-interaction/", srcPath: "auto-merge/contrib/plugins/web/opentelemetry-instrumentation-user-interaction/", bundleName: "otel-sndbx.instr-user-interaction", bundleNamespace: "opentelemetry.sandbox.instr.web.instrumentation-user-interaction", noWorkerTests: true, noNodeTests: true },
     { name: "@opentelemetry/instrumentation-long-task", destPath: "pkgs/instrumentations/web/long-task/", srcPath: "auto-merge/contrib/plugins/web/opentelemetry-instrumentation-long-task/", bundleName: "otel-sndbx.instr-long-task", bundleNamespace: "opentelemetry.sandbox.instr.web.instrumentation-long-task", noWorkerTests: true, noNodeTests: true },
     { name: "@opentelemetry/auto-instrumentations-web", destPath: "pkgs/instrumentations/web/auto/", srcPath: "auto-merge/contrib/metapackages/auto-instrumentations-web/", bundleName: "otel-sndbx.instr-auto", bundleNamespace: "opentelemetry.sandbox.instr.web.auto", noWorkerTests: true, noNodeTests: true },
-    { name: "@opentelemetry/opentelemetry-browser-detector", destPath: "pkgs/detectors/browser/", srcPath: "auto-merge/js/experimental/packages/opentelemetry-browser-detector/", bundleName: "otel-sndbx.detect-browser", bundleNamespace: "opentelemetry.sandbox.detector.browser", noWorkerTests: true }
+    { name: "@opentelemetry/opentelemetry-browser-detector", destPath: "pkgs/detectors/browser/", srcPath: "auto-merge/js/experimental/packages/opentelemetry-browser-detector/", bundleName: "otel-sndbx.detect-browser", bundleNamespace: "opentelemetry.sandbox.detector.browser", noWorkerTests: true },
+    { name: "@opentelemetry/api-logs", destPath: "pkgs/api-logs/", srcPath: "auto-merge/js/experimental/packages/api-logs/", bundleName: "otel-sndbx.api-logs", bundleNamespace: "opentelemetry.sandbox.api-logs", noWorkerTests: true, noBrowserTests: true },
+    { name: "@opentelemetry/api-events", destPath: "pkgs/api-events/", srcPath: "auto-merge/js/experimental/packages/api-events/", bundleName: "otel-sndbx.api-events", bundleNamespace: "opentelemetry.sandbox.api-events", noWorkerTests: true, noBrowserTests: true }
 ];
 
 /**
